Move src/axios.ts to src/utils/axios.ts
