+++ frontend/src/pages/_app.tsx
@@ -4,7 +4,7 @@
   @typescript-eslint/no-unsafe-member-access,
 */
 
-import '@/axios';
+import '@/utils/axios';
 import '@/tailwind.scss';
 import '@/global.scss';
