+++ frontend/src/pages/plugins/[name].tsx
@@ -5,7 +5,6 @@ import { SSRConfig, useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ParsedUrlQuery } from 'node:querystring';
 
-import { hubAPI } from '@/axios';
 import { ErrorMessage } from '@/components/ErrorMessage';
 import { PageMetadata } from '@/components/PageMetadata';
 import { PluginDetails } from '@/components/PluginDetails';
@@ -14,6 +13,7 @@ import { PluginStateProvider } from '@/context/plugin';
 import { PluginData } from '@/types';
 import { I18nNamespace } from '@/types/i18n';
 import { fetchRepoData, FetchRepoDataResult } from '@/utils';
+import { hubAPI } from '@/utils/axios';
 
 /**
  * Interface for parameters in URL.
